Add class selectors to html file and use selectors
Use Div tag to group the title and the images with text into two groups.
Add one more Div tag to nest the two Div in one container.
Add class selectors to to the Div tags, and use the selectors to declare
flex display.
Use flex-direction, align-items, and text-align to format the images and
text. Use flex-wrap to wrap the text.
Add paddings to space the text and images pairs from each other.
Add bottom margin to the title, and give the image text pairs a height
of 200px, to have give the a 200px width.
Add a gap of 52px between items.